Fine tuning
Fixed some formatting issues in files that was affecting key word searches. Fixed a build error with a corrupted build cache. The culprit was in the Android NDK compiler used by Unity's IL2CPP backend. Project is stable and I think finally done as far as features and accessibility. Now to add the rest of the information...NOTE: v1.3.15